updated state a& city
